requirements.txt
requests==2.32.3
beautifulsoup4==4.13.4
pandas==2.2.3
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-oauthlib==1.2.2
pytest==8.3.5
coverage==7.8.0